Sparse multi-threaded tests
* OpenCL Backend disabled
